Move interrupt_channel task into tasks/interrupts
