+++ core/tasks/interrupts/interrupt_channel.go
@@ -1,4 +1,4 @@
-package channels
+package interrupts
 
 import (
 	"context"
